pyproject.toml
[project]
name = "kgrypto"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "voynow", email = "voynow99@gmail.com"},
]
dependencies = ["polygon-api-client>=1.14.3", "pandas>=2.2.3", "requests>=2.32.3", "pydantic>=2.10.6", "polars>=1.21.0", "plotly>=6.0.0", "nbformat>=5.10.4"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
